requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
gunicorn==21.2.0
aiogram==3.3.0
google-genai==0.3.0
